Expense rating is displayed by stars, importance is displayed by color and value.
https://action.mindjet.com/task/14896530

+++ www/ExpensesAzure/Models/DataModels.cs
@@ -124,14 +124,14 @@ namespace SocialApps.Models
         {
             get
             {
-                return Rating != null ? Rating.ToString() : "";
+                return Rating != null ? ((short)Rating).ToString() : "";
             }
         }
         public string ImportanceString
         {
             get
             {
-                return Importance != null ? Importance.ToString() : "";
+                return Importance != null ? ((int)Importance).ToString() : "";
             }
         }
     }
